Support additional file-types and extensions when converting ADM to IAMF.
  - Pan objects to ambisonics, and extract the LFE when detected.
  - Input profile-version related flags may change in future CLs.

PiperOrigin-RevId: 706743391

+++ iamf/cli/adm_to_user_metadata/adm/bw64_reader.h
@@ -13,6 +13,7 @@
 #ifndef CLI_ADM_TO_USER_METADATA_ADM_BW64_READER_H_
 #define CLI_ADM_TO_USER_METADATA_ADM_BW64_READER_H_
 
+#include <cstddef>
 #include <cstdint>
 #include <istream>
 #include <string>
@@ -43,7 +44,7 @@ namespace adm_to_user_metadata {
 class Bw64Reader {
  public:
   struct ChunkInfo {
-    int32_t size, offset;
+    size_t size, offset;
   };
   typedef absl::flat_hash_map<std::string, ChunkInfo> ChunksOffsetMap;
